# perf: replace char comparison with array lookup