isNotEscapedSingleChar is a hotsport that was taking significant cpu time during encoding for string-heavy structs.
Replacing char comparison with a array lookup yields a significant performance improvement:

```
                 │   old.txt   │               new.txt
                 │   sec/op    │   sec/op     vs base
String/string-20   35.72n ± 3%   18.61n ± 2%  -47.89% (p=0.000 n=10)
```

Both variant are zero allocations.